Added and handled new custom exceptions, Edited UserService and CarService, migration V1 and UserRepository.

+++ src/main/java/com/hex/car_service_restful_app/services/CarServiceService.java
@@ -46,6 +46,9 @@ public class CarServiceService {
     }
 
     public void delete(String id) {
+        if (!carServiceRepository.existsById(Long.valueOf(id))){
+            throw new NotFoundException(String.format("Car service with id %s not found", id));
+        }
         carServiceRepository.deleteById(Long.valueOf(id));
     }
 }
